test: Move issue #503 regression tests into content::macros
Reserve parser/src/tests/asciidoc_lang/macros/autolinks.rs for tests derived
from the AsciiDoc language specification; the #503 regression tests live
alongside the code they exercise, in the inline_link test module.
